freqbench-helper: Dont forget to take SoC name input in int_freq_efficiency_graph()
SoC name is assigned when user gives input for fetching data.
But it becomes a problem when user uses local data to work with,
They get never asked to input SoC name and int_freq_efficiency_graph() fails to generate graph.

So add a input command in int_freq_efficiency_graph() to assign SoC name to a variable so that,
int_freq_efficiency_graph() wroks as needed.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>